pyproject.toml
[tool.poetry]
name = "adk-copilot"
version = "0.1.0"
description = "An AI copilot and expert assistant for developers using the Google Agent Development Kit (ADK)."
authors = ["Mohit Bhimrajka <mohitbhimrajka5@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-adk = "^1.0.0"
google-cloud-aiplatform = { extras = ["adk", "agent-engines", "evaluation"], version = "^1.93.0" }
google-cloud-storage = "^2.10.0"
pydantic = "^2.11.3"
python-dotenv = "^1.0.1"
sqlglot = "^26.10.1"
db-dtypes = "^1.4.2"
requests = "^2.31.0" # For future CRM/API tool integration
llama-index = "^0.12"
tabulate = "^0.9.0"
playwright = "^1.44.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"